resolve issue with undefined variable in Windows

+++ PHPCI/Model/Build/RemoteGitBuild.php
@@ -106,7 +106,9 @@ class RemoteGitBuild extends Build
 
         // Remove the key file and git wrapper:
         unlink($keyFile);
-        unlink($gitSshWrapper);
+        if (!IS_WIN) {
+            unlink($gitSshWrapper);
+        }
 
         return $success;
     }
